change result of multiply to long

+++ src/main/java/com/nobodyiam/test/Demo.java
@@ -10,7 +10,7 @@ public class Demo {
     return number1 + number2;
   }
 
-  public int multiply(int number1, int number2) {
+  public long multiply(int number1, int number2) {
     return number1 * number2;
   }
 }
